requirements.txt
# ===== Web Framework =====
fastapi
uvicorn[standard]

# ===== Auth / Security =====!
python-dotenv
python-multipart
itsdangerous

# ===== Google API =====
google-auth
google-auth-oauthlib
google-api-python-client

# ===== Data Handling =====
pandas
openpyxl
xlrd
python-dateutil
psycopg2

# ===== AI / LLM =====
openai

# ===== HTTP =====
requests

# ===== DB (지금 or 향후 대비) =====
sqlalchemy
supabase



# ===== batch ======
tzdata>=2024.1

# ===== 알림 시스템 =====
redis
python-socketio
python-engineio
firebase-admin